+++ android-plugin/src/main/java/io/creedengo/xml/Xml.java
@@ -1,0 +1,27 @@
+/*
+ * Creedengo Android Java plugin - Provides rules to reduce the environmental footprint of your Android applications
+ * Copyright © 2020 Green Code Initiative (contact@green-code-initiative.org)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package io.creedengo.xml;
+
+public interface Xml { //NOSONAR - we use the interface for constant without inheriting it
+    String REPOSITORY_NAME = "Creedengo (Android XML)";
+    String KEY = "xml";
+    String REPOSITORY_KEY = "creedengo-android-xml";
+    String RULES_SPECIFICATIONS_XML_PATH = "io/creedengo/rules/xml";
+    String PROFILE_NAME = "Creedengo (Android XML)";
+    String PROFILE_PATH = "io/creedengo/android/xml/creedengo_xml_profile.json";
+}