+++ android-plugin/src/main/java/io/creedengo/xml/checks/idleness/KeepScreenOnXmlRule.java
@@ -1,0 +1,39 @@
+/*
+ * Creedengo Android Java plugin - Provides rules to reduce the environmental footprint of your Android applications
+ * Copyright © 2020 Green Code Initiative (contact@green-code-initiative.org)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package io.creedengo.xml.checks.idleness;
+
+import io.creedengo.xml.checks.XPathSimpleCheck;
+import org.sonar.check.Rule;
+import org.sonarsource.analyzer.commons.annotations.DeprecatedRuleKey;
+
+@DeprecatedRuleKey(repositoryKey = "creedengo-android-xml", ruleKey = "EC541")
+@Rule(key = "GCI541")
+@DeprecatedRuleKey(repositoryKey = "ecocode-android-xml", ruleKey = "EIDL003")
+public class KeepScreenOnXmlRule extends XPathSimpleCheck {
+
+    private static final String KEEP_SCREEN_ON_XML_ATTRIBUTE = "//*[contains(@keepScreenOn, 'true')]/@keepScreenOn";
+    private static final String ERROR_MESSAGE = "Keeping the screen on should be avoided to avoid draining the battery.";
+
+    protected String getMessage() {
+        return ERROR_MESSAGE;
+    }
+
+    protected String getXPathExpressionString() {
+        return KEEP_SCREEN_ON_XML_ATTRIBUTE;
+    }
+}