+++ android-plugin/src/test/files/social/privacy/GoogleTrackerFirebaseCheck.java
@@ -1,6 +1,6 @@
 /*
- * ecoCode Android plugin - Provides rules to reduce the environmental footprint of your Android applications
- * Copyright © 2020 Green Code Initiative (contact@ecocode.io)
+ * Creedengo Android Java plugin - Provides rules to reduce the environmental footprint of your Android applications
+ * Copyright © 2020 Green Code Initiative (contact@green-code-initiative.org)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,6 +18,6 @@
 import com.google.android.gms.analytics.Tracker; // Noncompliant {{Using com.google.android.gms.analytics.* is a potential threat for privacy.}}
 import com.google.android.gms.Test;
 import com.google.firebase.analytics.FirebaseAnalytics; // Noncompliant {{Using com.google.firebase.analytics.* is a potential threat for privacy.}}
-import com.io.ecocode.MaClasse;
+import com.io.creedengo.MaClasse;
 public class Test {
 }